Set standard lifetime for add2cahce objects to 1 second
A timeout cann occur, if the add2cache request is sent just before the ageing process takes place. It deletes the add2chache interest and the content will be discarded as unsolicited. This is a workaround, a better fix would be to have a flag here that indicates that add2cache interests should not be deleted by ageing since they are deleted on completion of add2cache.